r2engine: got rid of material handles in the GPUModelRef

+++ r2engine/src/r2/Render/Renderer/Shader.h
@@ -10,7 +10,7 @@
 namespace r2::draw
 {
 
-    using ShaderHandle = u32;
+    using ShaderHandle = u64;
     using ShaderName = u64;
     static const ShaderHandle InvalidShader;
